importccl: move export parquet tests to new file
Release note: None
